Разбиваем код на модули
